Move Fixtures.kt to api.test directory
