+++ Drop-In/src/sharedTest/java/com.braintreepayments.api.test/Fixtures.kt
@@ -1,4 +1,4 @@
-package com.braintreepayments.api
+package com.braintreepayments.api.test
 
 object Fixtures {
