+++ Drop-In/src/test/java/com/braintreepayments/api/dropin/DropInResultUnitTest.java
@@ -7,7 +7,7 @@ import androidx.fragment.app.FragmentManager;
 
 import com.braintreepayments.api.BraintreeFragment;
 import com.braintreepayments.api.ConfigurationManagerTestUtils;
-import com.braintreepayments.api.Fixtures;
+import com.braintreepayments.api.test.Fixtures;
 import com.braintreepayments.api.GooglePayment;
 import com.braintreepayments.api.dropin.utils.PaymentMethodType;
 import com.braintreepayments.api.exceptions.InvalidArgumentException;
