+++ Drop-In/src/test/java/com/braintreepayments/api/dropin/utils/PaymentMethodTypeUnitTest.java
@@ -1,6 +1,6 @@
 package com.braintreepayments.api.dropin.utils;
 
-import com.braintreepayments.api.Fixtures;
+import com.braintreepayments.api.test.Fixtures;
 import com.braintreepayments.api.dropin.R;
 import com.braintreepayments.api.models.CardNonce;
 import com.braintreepayments.api.models.PayPalAccountNonce;
