+++ Drop-In/src/test/java/com/braintreepayments/api/dropin/view/AddCardViewUnitTest.java
@@ -5,7 +5,7 @@ import android.text.SpannableString;
 import android.view.View;
 import android.widget.Button;
 
-import com.braintreepayments.api.Fixtures;
+import com.braintreepayments.api.test.Fixtures;
 import com.braintreepayments.api.dropin.R;
 import com.braintreepayments.api.dropin.interfaces.AddPaymentUpdateListener;
 import com.braintreepayments.api.dropin.utils.PaymentMethodType;
